[refactor-3] created backdrops components
