+++ src/components/Dashboard/Backdrop/AddPosition/style.css
@@ -1,0 +1,11 @@
+.backdropWrapper {
+  display: block !important;
+}
+
+.modal-dialog.backdropModal {
+  max-width: 60%;
+}
+
+form label {
+  font-weight: bold;
+}
