+++ src/components/Dashboard/Backdrop/ShowPosition/style.css
@@ -1,0 +1,24 @@
+.backdropWrapper {
+  display: block !important;
+}
+
+.modal-dialog.backdropModal {
+  max-width: 60%;
+}
+
+form label {
+  font-weight: bold;
+}
+
+.displayForm {
+  border: 1.5px solid #1d9624;
+}
+.displayForm:disabled {
+  background-color: transparent;
+  border: none;
+}
+
+.hideButton {
+  display: none;
+  visibility: hidden;
+}
