+++ src/components/Dashboard/Homepage/style.css
@@ -26,28 +26,3 @@
   border-radius: 1rem;
   padding: 0.5rem;
 }
-
-.backdropWrapper {
-  display: block !important;
-}
-
-.modal-dialog.backdropModal {
-  max-width: 60%;
-}
-
-form label {
-  font-weight: bold;
-}
-
-.displayForm {
-  border: 1.5px solid #1d9624;
-}
-.displayForm:disabled {
-  background-color: transparent;
-  border: none;
-}
-
-.hideButton {
-  display: none;
-  visibility: hidden;
-}
